requirements.txt
PyYAML
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
pdfplumber
pytesseract
pillow
requests
python-dotenv
